refactor(core,copilot): Reduce duplication across handlers and tools
Six small cleanups surfaced by code review, none affecting the public
plugin surface beyond making BlockEditorHandlerBase a public extension
point for third-party block-shaped editors:

- Extract BlockEditorHandlerBase as the shared base for BlockList /
  BlockGrid handlers. The two concrete handlers now hold only the editor
  schema alias, the layout key, and a layout-entry builder; everything
  else (add/remove/move/clear, expose, item-property accessors,
  content-type and editor-alias resolution) lives in one place.
- Picker handlers' TryGetGuid now delegates to BlockEnvelopeOps.GetGuid
  (promoted from private to internal) — same body in three places becomes
  one.
- Move parsePath / readVariant from add-item.api.ts to
  internal/path-args.ts so the four sibling tools no longer cross-import
  from each other (a leaky API).
- Replace the local uuidv4() in entity-adapter/value-preparation.ts with
  crypto.randomUUID(), matching the convention already used in 8+ places
  across the client codebase.
- Drop the redundant upfront path-shape validation loop in the dispatcher.
  Per-segment shape is enforced by the JSON converter at deserialization;
  the descent's casts surface any other mismatch as a structured
  InvalidPath. The remaining check (leaf must be a property alias) is the
  only constraint the converter alone can't express.
- Trim a handful of narrative comments that restated the surrounding code
  (file:line markers like "Resolve handler", "Snapshot the frame so we
  can ascend later", etc.). Kept the WHY comments.

Build and test: 794 unit tests pass; all five frontend packages build
clean.
